+++ examples/aws-docker-image-simple/variables.tf
@@ -3,11 +3,6 @@ variable "domain" {
   type        = string
 }
 
-variable "email" {
-  description = "Email address used when registering certificates with Let's Encrypt."
-  type        = string
-}
-
 variable "zone_id" {
   description = "Route53 Zone ID."
   type        = string