+++ main.tf
@@ -11,6 +11,8 @@ locals {
     for k, v in var.container : k => v if !contains(["image", "container_name", "command"], k)
   }
 
+  # image can be either var.container.image (as is documented) or
+  # var.image (as previously implemented)
   image = lookup(var.container, "image", var.image)
 
   # Define environment variables that will be written to `/var/app/.env` and
@@ -21,8 +23,6 @@ locals {
 
   environment = merge({
     DOMAIN                     = var.domain
-    LETSENCRYPT_EMAIL          = var.email
-    LETSENCRYPT_SERVER         = var.letsencrypt_staging ? "https://acme-staging-v02.api.letsencrypt.org/directory" : null
     COMPOSE_DOCKER_IMAGE       = "docker/compose"
     COMPOSE_DOCKER_TAG         = "1.29.2"
     IMAGE_NAME                 = try(split(":", local.image)[0], null)